==> step4-llm_select/0.data_process.ipynb <==
import pandas as pd
import os

dataset_name = 'eu'
df = pd.read_excel(f'../step3-bert_recall/CuBERT/results/{dataset_name}/2e-05_32_0.8/{dataset_name}.xlsx')
for seed in range(5, 10):
    if not os.path.exists(f'data/{dataset_name}/seed{seed}/input0'):
        os.makedirs(f'data/{dataset_name}/seed{seed}/input0')
    df.to_csv(f'data/{dataset_name}/seed{seed}/input0/task_candidate.csv', index=None, sep='\t')
df
