==> step4-llm_select/3.result.ipynb <==
import pandas as pd
import os

dataset_name = 'eu'

for seed in range(6):
    ans = []
    for turn in range(5):
        if os.path.exists(f'results/{dataset_name}/seed{seed}/result{turn}.parquet'):
            temp = pd.read_parquet(f'results/{dataset_name}/seed{seed}/result{turn}.parquet')
            temp.columns = ['job_title', 'job_description', f'task_seed{seed}']
            temp[f'task_seed{seed}'] = temp[f'task_seed{seed}'].apply(lambda x: [x[i] for i in range(len(x)) if x[i] not in x[:i]])
            ans.append(temp)
    if len(ans) == 0:
        continue
    ans = pd.concat(ans)
    if seed == 0:
        final_task = ans
    else:
        final_task = pd.merge(final_task, ans, on=['job_title', 'job_description'])
final_task

origin_df = pd.read_csv(f'../step3-bert_recall/CuBERT/results/{dataset_name}/2e-05_32_0.8/{dataset_name}.csv', sep='\t')
origin_df['sim'] = origin_df.apply(lambda x: sum([x[f"task{i}_sim"] for i in range(100)]), axis=1)
origin_df = origin_df[['job_title', 'job_description', 'sim']]
final_task = pd.merge(final_task, origin_df, on=['job_title', 'job_description'])
final_task

from utlis import aggregate_expert_opinions
final_task['task_top10'] = final_task.apply(lambda x: aggregate_expert_opinions([x[f'task_seed{i}'] for i in range(6) if f'task_seed{i}' in x], top_num=10)[0], axis=1)
final_task['task'] = final_task['task_top10'].apply(lambda x: x[0])
final_task['score'] = final_task.apply(lambda x: aggregate_expert_opinions([x[f'task_seed{i}'] for i in range(6) if f'task_seed{i}' in x], top_num=1)[1][0], axis=1)
final_ans = final_task[['job_title', 'job_description', 'task', 'sim', 'score', 'task_top10']]
final_ans = final_ans.sort_values(['score', 'sim'], ascending=[False, False])
final_ans.to_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx', index=None)
final_ans
