==> step4-llm_select/4.gpt_select.ipynb <==
import pandas as pd
thred = 20
dataset_name = 'en'
final_ans = pd.read_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx')
sub_ans = final_ans[(final_ans['sim']>0.9) & (final_ans['score']>=thred)]
ans_group = sub_ans.groupby('task').agg(list)[['job_description']]
ans_group['len'] = ans_group['job_description'].apply(len)
ans_group = ans_group.sort_values('len', ascending=False)
annotated_ans = ans_group[ans_group['len']>=10].reset_index()
annotated_ans = annotated_ans.explode('job_description')
origin_df = pd.read_csv(f'data/{dataset_name}/seed0/input0/task_candidate.csv', sep='\t')
annotated_ans = pd.merge(annotated_ans, origin_df, on=['job_description'])
annotated_ans.to_csv(f'annotated_gpt/inputs/{dataset_name}.csv', sep='\t', index=None)
annotated_ans.to_excel(f'annotated_gpt/inputs/{dataset_name}.xlsx', index=None)
annotated_ans

# from utlis import gen
# annotated_ans['prompt'] = annotated_ans.apply(lambda x: gen(x, dataset_name), axis=1)
# import os
# if not os.path.exists(f"annotated_gpt/inputs/{dataset_name}"):
#     os.makedirs(f"annotated_gpt/inputs/{dataset_name}")
# if not os.path.exists(f"annotated_gpt/outputs/{dataset_name}"):
#     os.makedirs(f"annotated_gpt/outputs/{dataset_name}")
# for i in range(len(annotated_ans)):
#     with open(f"annotated_gpt/inputs/{dataset_name}/{i}.txt", 'w') as w:
#         w.write(annotated_ans['prompt'][i])
    # with open(f"annotated_gpt/outputs/{dataset_name}/{i}.txt", 'w') as w:
    #     w.write('')
