==> step4-llm_select/5.gpt_answer.ipynb <==
import pandas as pd
import os
import re
from tqdm import tqdm
dataset_name = 'en'
input_dir = f"annotated_gpt/inputs/{dataset_name}"
output_dir = f"annotated_gpt/outputs/{dataset_name}"

all_data_task_topk = pd.read_csv(f'annotated_gpt/inputs/{dataset_name}.csv', sep='\t').reset_index()


final_ans = {'index':[], 'gpt_task': []}
for i in tqdm(os.listdir(output_dir)):
    idx = int(re.findall(r'\d+', i)[0])
    text = open(f"{output_dir}/{idx}.txt", 'r').read()
    content = text
    text = open(f"{output_dir}/{idx}.txt", 'r').read()
    ans = re.findall(r'\d\.+(.*?)\n', content)
    candidates = [all_data_task_topk[f'task{_}'][idx].strip(' ') for _ in range(100)]
    results = []
    # rgex = '|'.join([_ for _ in  candidates])
    # results = re.findall(rf"{rgex}", content)
    for line in content.split('\n'):
        for can in candidates:
            if can in line:
                results.append(can)
                break
    if len(results) != 10:
        continue
    final_ans['gpt_task'].append(results)
    final_ans['index'].append(idx)
final_ans = pd.DataFrame(final_ans)
# 处理输出文件
final_ans = pd.merge(all_data_task_topk[['index', 'job_title', 'job_description', 'task']], final_ans, on=['index'])
final_ans.to_csv(f'annotated_gpt/outputs/{dataset_name}.csv', index=None)
final_ans

# final_ans = final_ans[final_ans['index']>=147]

final_ans['gpt_task'] = final_ans['gpt_task'].apply(lambda x: x[0])
final_group = final_ans.groupby('gpt_task').agg(list)[['job_description']]
final_group['len'] = final_group['job_description'].apply(len)
final_group = final_group.sort_values('len', ascending=False)
final_group = final_group.explode('job_description').reset_index()[['gpt_task', 'job_description', 'len']]
final_group = pd.merge(final_group, final_ans, on=['gpt_task', 'job_description'])[['job_title', 'job_description', 'gpt_task', 'task', 'len']]
if not os.path.exists(f'annotated_gpt/results'):
    os.makedirs(f'annotated_gpt/results')
final_group.to_excel(f'annotated_gpt/results/{dataset_name}.xlsx', index=None)
final_group

# import pandas as pd
# df1 = pd.read_excel('annotated_gpt/results/en.xlsx')
# df2 = pd.read_excel('annotated_gpt/results/en-annotated.xlsx')
# pd.merge(df1, df2, how='left').to_excel('annotated_gpt/results/en-annotated1.xlsx', index=None)
