==> step4-llm_select/6.final_data.ipynb <==
import pandas as pd
dataset_name = 'en'
df = pd.read_excel(f'annotated_gpt/results/{dataset_name}-annotated.xlsx')
df = df[df['annotated'].isna()].drop(['len', 'annotated', 'task'], axis=1)
df.columns = ['job_title', 'job_description', 'task']
df.to_excel(f'results/{dataset_name}-final.xlsx', index=None)
df

gpt_df = pd.read_csv(f'annotated_gpt/outputs/{dataset_name}.csv').drop(['index', 'task'], axis=1)
gpt_df = pd.merge(df, gpt_df, on=['job_title', 'job_description'])
llm_df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx')[['job_title', 'job_description', 'task_top5']]
llm_df.columns = ['job_title','job_description', 'llm_task']
llm_df = pd.merge(gpt_df, llm_df, on=['job_title', 'job_description'])
bert_df = pd.read_csv(f'data/{dataset_name}/seed0/input0/task_candidate.csv', sep='\t')
bert_df['bert_task'] = bert_df.apply(lambda x: [x[f"task{i}"] for i in range(100)], axis=1)
bert_df = bert_df[['job_title', 'job_description', 'bert_task']]
bert_df = pd.merge(llm_df, bert_df, on=['job_title', 'job_description'])
bert_df.to_csv(f'results/{dataset_name}/{dataset_name}_gpt-llm-bert.csv',index=None)
bert_df
