==> step4-llm_select/7.supplement_data.ipynb <==
# import pandas as pd
# datastatics = {}
# thred_score = {"en": 25, "eu": 25, "jp": 20}
# for dataset_name in ['eu', 'en', 'jp']:
#     datastatics[dataset_name] = {}
#     df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx')
#     final_df = pd.read_excel(f'results/{dataset_name}-final.xlsx')
#     sub_df = df[~df['job_description'].isin(final_df['job_description'])]
#     sub_df = pd.concat([sub_df, final_df])
#     datastatics[dataset_name]['origin_length'] = len(sub_df)
#     # sub_df.to_excel(f'results/{dataset_name}-llm.xlsx', index=None)
#     sub_df = sub_df[sub_df['score']>=thred_score[dataset_name]]
#     sub_df_group = sub_df.groupby('task').agg(list)[['job_description']]
#     sub_df_group['len'] = sub_df_group['job_description'].apply(len)
#     sub_df_group = sub_df_group.sort_values('len', ascending=False)
#     for thred_len in [0, 5, 10, 15, 20, 25, 40]:
#         all_df = sub_df_group[sub_df_group['len']>=thred_len].explode('job_description')[['job_description']].reset_index()
#         if thred_len == 0:
#             all_df.to_excel(f'results/{dataset_name}/{dataset_name}_all.xlsx', index=None)
#             datastatics[dataset_name]['length'] = len(all_df)
#         datastatics[dataset_name][thred_len] = all_df['task'].nunique()
# pd.DataFrame(datastatics)

import pandas as pd
new_df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_all.xlsx')
df_ch = pd.read_excel(f'results/{dataset_name}/{dataset_name}_all_ch.xlsx')
ch2jp = {v:i for i,v in zip(new_df['task'], df_ch['任务'])}

df_group = df_ch.groupby('任务').agg(list)
df_group['len'] = df_group['职位描述'].apply(len)
df_group = df_group.sort_values('len', ascending=False).reset_index()
df_group

key_word = '检查'
list_set = df_ch[df_ch['任务'].apply(lambda x: key_word in x)]['任务'].unique()
temp = df_group[df_group['任务'].isin(list_set)][['任务', 'len']].reset_index(drop=True)
temp.to_csv('temp.csv',index=None)
content = open('temp.csv', 'r').read() + "\n\n给定上述表格，我希望能对一些任务进行合并，尤其是将数量较少的task合并到数量较高的task，请返回json文件，关键字是需要被合并的任务，字段内容是合并到的task"
with open('temp.csv', 'w') as w:
    w.write(content)

import json
import os
merge_map_ch = {}
for file in os.listdir(f'results/{dataset_name}/merge_json'):
    if file.endswith('json'):
        merge_map_ch.update(json.load(open(f'results/{dataset_name}/merge_json/{file}', 'r')))
merge_map = {ch2jp[i]:ch2jp[v] for i,v in merge_map_ch.items()}
new_df['task'] = new_df['task'].apply(lambda x: merge_map[x] if x in merge_map else x)

import pandas as pd
datastatics = {}
thred_score = {"en": 25, "eu": 25, "jp": 20}
dataset_name = 'jp'

datastatics[dataset_name] = {}
df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx')
final_df = pd.read_excel(f'results/{dataset_name}-final.xlsx')
sub_df = df[~df['job_description'].isin(final_df['job_description'])]
sub_df = pd.concat([sub_df, final_df])

sub_df = pd.merge(sub_df.drop('task', axis=1), new_df, on=['job_description']) 


datastatics[dataset_name]['origin_length'] = len(sub_df)
# sub_df.to_excel(f'results/{dataset_name}-llm.xlsx', index=None)
sub_df = sub_df[sub_df['score']>=thred_score[dataset_name]]


sub_df_group = sub_df.groupby('task').agg(list)[['job_description']]
sub_df_group['len'] = sub_df_group['job_description'].apply(len)
sub_df_group = sub_df_group.sort_values('len', ascending=False)
for thred_len in [0, 5, 10, 15, 20, 25, 40]:
    all_df = sub_df_group[sub_df_group['len']>=thred_len].explode('job_description')[['job_description']].reset_index()
    if thred_len == 0:
        all_df.to_excel(f'results/{dataset_name}/{dataset_name}_all.xlsx', index=None)
        datastatics[dataset_name]['length'] = len(all_df)
    datastatics[dataset_name][thred_len] = all_df['task'].nunique()
    
pd.DataFrame(datastatics)

import pandas as pd
metrics = {}
for dataset_name in ['en', 'eu', 'jp']:
    metrics[dataset_name] = {}
    df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_all.xlsx')
    df_group = df.groupby('task').agg(list)
    df_group['len'] = df_group['job_description'].apply(len)
    df_group = df_group.sort_values('len', ascending=False)
    df_group = df_group[df_group['len']>-10]
    metrics[dataset_name]['class_num>20'] = len(df_group[df_group['len']>=20])
    metrics[dataset_name]['20>class_num>=10'] = len(df_group[(df_group['len']<20) & (df_group['len']>=10)])
    df = df_group[df_group['len']>=10].explode('job_description').reset_index()
    df.to_excel(f"results/{dataset_name}/{dataset_name}_filter.xlsx", index=None)
    metrics[dataset_name]['instance_num'] = len(df)
pd.DataFrame(metrics)

import pandas as pd
for dataset_name in ['en', 'eu', 'jp']:
    df = pd.read_excel(f"results/{dataset_name}/{dataset_name}_filter.xlsx")
    llm_df = pd.read_excel(f'results/{dataset_name}/{dataset_name}_results.xlsx')[['job_title', 'job_description', 'task_top10']]
    llm_df.columns = ['job_title','job_description', 'llm_task']
    llm_df = pd.merge(df, llm_df, on=['job_description'])
    bert_df = pd.read_csv(f'data/{dataset_name}/seed0/input0/task_candidate.csv', sep='\t')
    bert_df['bert_task'] = bert_df.apply(lambda x: [x[f"task{i}"] for i in range(100)], axis=1)
    bert_df = bert_df[['job_title', 'job_description',  'bert_task']]
    final_df = pd.merge(llm_df, bert_df, on=['job_title', 'job_description'])
    final_df = final_df[['job_title', 'job_description', 'task', 'llm_task', 'bert_task']]
    final_df.to_csv(f'results/{dataset_name}/{dataset_name}_filter.csv')
final_df
