==> step4-llm_select/8.data_statics.ipynb <==
import pandas as pd
for dataset_name in ['eu', 'en', 'jp']:
    df = pd.read_csv((f'results/{dataset_name}/{dataset_name}_filter.csv'))
    print(dataset_name, df['job_title'].nunique(), df['job_description'].nunique())
